refactor: fixed로 변경, 컴포넌트 파일 분리, 공통 스타일 지정
